Compenetize menu avatar and add to map page
