+++ src/client/screens/HomeScreen.tsx
@@ -1,11 +1,12 @@
 /** @jsx jsx */
-import { Button as MenuButton, Wrapper, Menu, MenuItem } from "react-aria-menubutton";
-import Avatar from "react-avatar";
 import React, { useEffect, useState } from "react";
+import { projectsFetch } from "../actions/projects";
+import { userFetch } from "../actions/user";
 import { connect } from "react-redux";
 import { Link, RouteComponentProps } from "react-router-dom";
-import * as H from "history";
 import Icon from "../components/Icon";
+import AvatarMenu from "../components/AvatarMenu";
+import store from "../store";
 import {
   Alert,
   Box,
@@ -21,30 +22,17 @@ import {
 import { ReactComponent as Logo } from "../media/logos/logo.svg";
 import { ReactComponent as NoMapsIllustration } from "../media/no-maps-illustration.svg";
 
-import { resetState } from "../actions/root";
-import { projectsFetch } from "../actions/projects";
-import { userFetch } from "../actions/user";
 import { resendConfirmationEmail } from "../api";
-import { clearJWT, getJWT } from "../jwt";
+import { getJWT } from "../jwt";
 import { State } from "../reducers";
 import { ProjectsState } from "../reducers/projects";
 import { UserState } from "../reducers/user";
 import { WriteResource } from "../resource";
-import store from "../store";
 
 interface StateProps {
   readonly projects: ProjectsState;
   readonly user: UserState;
 }
-
-enum UserMenuKeys {
-  Logout = "logout"
-}
-
-const logout = () => {
-  clearJWT();
-  store.dispatch(resetState());
-};
 
 const style: ThemeUIStyleObject = {
   header: {
@@ -62,68 +50,6 @@ const style: ThemeUIStyleObject = {
     "&:focus": {
       outline: "none",
       boxShadow: "focus"
-    }
-  },
-  avatar: {
-    fontFamily: "heading",
-    cursor: "pointer",
-    ".sb-avatar__text": {
-      "&:hover": {
-        backgroundColor: "#395c78 !important"
-      },
-      "&:active": {
-        backgroundColor: "#2c485e !important"
-      }
-    }
-  },
-  menuButton: {
-    display: "flex",
-    alignItems: "center",
-    bg: "transparent",
-    p: 1,
-    borderRadius: "small",
-    "&:focus": {
-      outline: "none",
-      boxShadow: "focus"
-    }
-  },
-  menu: {
-    width: "150px",
-    position: "absolute",
-    mt: 2,
-    right: 2,
-    bg: "muted",
-    py: 1,
-    px: 1,
-    border: "1px solid",
-    borderColor: "gray.2",
-    boxShadow: "small",
-    borderRadius: "small"
-  },
-  menuList: {
-    p: "0",
-    m: "0",
-    listStyleType: "none"
-  },
-  menuListItem: {
-    borderRadius: "small",
-    py: 1,
-    px: 2,
-    "&:hover:not([disabled])": {
-      bg: "gray.0",
-      cursor: "pointer"
-    },
-    "&[disabled]": {
-      color: "gray.3",
-      cursor: "not-allowed"
-    },
-    "&:focus": {
-      bg: "gray.0",
-      outline: "none",
-      boxShadow: "focus"
-    },
-    "&:active": {
-      bg: "gray.1"
     }
   },
   projectRow: {
@@ -208,41 +134,7 @@ const HomeScreen = ({ projects, user, history }: StateProps & RouteComponentProp
             <Logo sx={{ width: "15rem" }} />
           </Link>
         </Heading>
-        {!isLoggedIn ? (
-          <React.Fragment>
-            <Link to="/login" sx={{ p: 2 }}>
-              Login
-            </Link>{" "}
-            <Link to="/register" sx={{ p: 2 }}>
-              Register
-            </Link>
-          </React.Fragment>
-        ) : "resource" in user ? (
-          <Wrapper onSelection={handleSelection(history)}>
-            <MenuButton sx={style.menuButton}>
-              <Avatar
-                name={user.resource.name}
-                round={true}
-                size={"2.5rem"}
-                color={"#2c485e"}
-                maxInitials={3}
-                sx={style.avatar}
-              />
-              <div sx={{ ml: 2 }}>
-                <Icon name="chevron-down" />
-              </div>
-            </MenuButton>
-            <Menu sx={style.menu}>
-              <ul sx={style.menuList}>
-                <li key={UserMenuKeys.Logout}>
-                  <MenuItem value={UserMenuKeys.Logout} sx={style.menuListItem}>
-                    Logout
-                  </MenuItem>
-                </li>
-              </ul>
-            </Menu>
-          </Wrapper>
-        ) : null}
+        <AvatarMenu user={user} history={history} />
       </Flex>
       <Flex
         as="main"
@@ -317,14 +209,6 @@ const HomeScreen = ({ projects, user, history }: StateProps & RouteComponentProp
       </Flex>
     </Flex>
   );
-};
-
-const handleSelection = (history: H.History) => (key: string | number) => {
-  // eslint-disable-next-line
-  if (key === UserMenuKeys.Logout) {
-    logout();
-    history.push("/login");
-  }
 };
 
 function mapStateToProps(state: State): StateProps {
